#14.2 ~ #16.3
남은 과제 : 애니메이션 만들기

13:19
#14.3 Unity RPG Tutorial - Looting items

26:31
#14.4 Unity RPG Tutorial - Loot pages

14:31
#14.5 Unity RPG Tutorial - Loot interaction

30:46
#14.6 Unity RPG Tutorial - Loot table

15:55
#15.0 Unity RPG Tutorial - Character panel

27:13
#15.1 Unity RPG Tutorial - Equipping armor

24:37
#15.2 Unity RPG Tutorial - Click to equip

12:24
#15.3 Unity RPG Tutorial - Char tooltip

22:10
#15.4  Unity RPG Tutorial - Dequipping gear

21:56
#16.0 Unity RPG Tutorial - Gear representation

22:05
#16.1 Unity RPG Tutorial - Gear animations

24:46
#16.2 Unity RPG Tutorial - Equipping armor

11:51
#16.3 Unity RPG Tutorial - Weapon socket

+++ Assets/Scripts/UIRelated/HandScript.cs
@@ -26,12 +26,21 @@ public class HandScript : MonoBehaviour
         icon = GetComponent<Image>();
     }
 
-    void Update()
+
+    private void Update()
     {
+        // 마우스를 따라 아이콘이 이동한다.
         icon.transform.position = Input.mousePosition + offset;
 
-        DeleteItem();
+        // 마우스 왼쪽클릭이 되었다. 그리고
+        // 왼쪽클릭한 대상이 UI 가 아니다. 그리고
+        // MyMoveable 이 null 이 아니다.
+        if (Input.GetMouseButton(0) && !EventSystem.current.IsPointerOverGameObject() && Dragable != null)
+        {
+            DeleteItem();
+        }
     }
+    
 
     //SpellButton 스크립트에서 실행 -> 스킬창에 있는 스킬버튼
     public void TakeMoveable(IMoveable moveable)
@@ -56,18 +65,20 @@ public class HandScript : MonoBehaviour
     {
         Dragable = null;
         icon.color = new Color(0, 0, 0, 0);
+        InventoryScript.instance.ChoosedSlot = null;
     }
 
-    private void DeleteItem()
+    public void DeleteItem()
     {
-        if (Input.GetMouseButton(0) && !EventSystem.current.IsPointerOverGameObject() && instance.Dragable != null)
+
+        // Dragable인터페이스를 포함한 아이템이고, 선택가능한 슬롯이면
+        if (Dragable is Item && InventoryScript.instance.ChoosedSlot != null)
         {
-            if (Dragable is Item && InventoryScript.instance.ChoosedSlot != null)
-            {
-                (Dragable as Item).MySlot.Clear();
-            }
-            Drop();
-            InventoryScript.instance.ChoosedSlot = null;
+            // 모두 만족하면 해당 슬롯의 모든 아이템을 삭제한다.
+            (Dragable as Item).MySlot.Clear();
         }
+        // HandScript 의 아이콘 정보 초기화
+        Drop();
+        InventoryScript.instance.ChoosedSlot = null;
     }
 }
